Make MiniMax-H3 modular only
