+++ src/diffusers/__init__.py
@@ -725,8 +725,6 @@ else:
             "MarigoldDepthPipeline",
             "MarigoldIntrinsicsPipeline",
             "MarigoldNormalsPipeline",
-            "MiniMaxH3Pipeline",
-            "MiniMaxH3Ref2VAPipeline",
             "MochiPipeline",
             "MotifVideoImage2VideoPipeline",
             "MotifVideoPipeline",
@@ -1551,8 +1549,6 @@ if TYPE_CHECKING or DIFFUSERS_SLOW_IMPORT:
             MarigoldDepthPipeline,
             MarigoldIntrinsicsPipeline,
             MarigoldNormalsPipeline,
-            MiniMaxH3Pipeline,
-            MiniMaxH3Ref2VAPipeline,
             MochiPipeline,
             MotifVideoImage2VideoPipeline,
             MotifVideoPipeline,
