--- src/diffusers/pipelines/minimax_h3/__init__.py
@@ -1,54 +1,0 @@
-from typing import TYPE_CHECKING
-
-from ...utils import (
-    DIFFUSERS_SLOW_IMPORT,
-    OptionalDependencyNotAvailable,
-    _LazyModule,
-    get_objects_from_module,
-    is_torch_available,
-    is_transformers_available,
-)
-
-
-_dummy_objects = {}
-_import_structure = {}
-
-
-try:
-    if not (is_transformers_available() and is_torch_available()):
-        raise OptionalDependencyNotAvailable()
-except OptionalDependencyNotAvailable:
-    from ...utils import dummy_torch_and_transformers_objects  # noqa F403
-
-    _dummy_objects.update(get_objects_from_module(dummy_torch_and_transformers_objects))
-else:
-    _import_structure["packing_ref2va"] = ["MiniMaxH3Reference"]
-    _import_structure["pipeline_minimax_h3"] = ["MiniMaxH3Pipeline"]
-    _import_structure["pipeline_minimax_h3_ref2va"] = ["MiniMaxH3Ref2VAPipeline"]
-    _import_structure["pipeline_output"] = ["MiniMaxH3PipelineOutput"]
-
-if TYPE_CHECKING or DIFFUSERS_SLOW_IMPORT:
-    try:
-        if not (is_transformers_available() and is_torch_available()):
-            raise OptionalDependencyNotAvailable()
-
-    except OptionalDependencyNotAvailable:
-        from ...utils.dummy_torch_and_transformers_objects import *
-    else:
-        from .packing_ref2va import MiniMaxH3Reference
-        from .pipeline_minimax_h3 import MiniMaxH3Pipeline
-        from .pipeline_minimax_h3_ref2va import MiniMaxH3Ref2VAPipeline
-        from .pipeline_output import MiniMaxH3PipelineOutput
-
-else:
-    import sys
-
-    sys.modules[__name__] = _LazyModule(
-        __name__,
-        globals()["__file__"],
-        _import_structure,
-        module_spec=__spec__,
-    )
-
-    for name, value in _dummy_objects.items():
-        setattr(sys.modules[__name__], name, value)
